[Call,Access 1]분리 완료

+++ src/main/java/net/sourceforge/pmd/lang/java/rule/stanly/metrics/PackageSetMetric.java
@@ -60,7 +60,7 @@ public class PackageSetMetric{
 		return Fat;
 	}
 	public void setFat(float fAT) {
-		Fat = fAT;
+		Fat = (int) fAT;
 	}
 	public float getTangled() {
 		return Tangled;
